Library for coaps server
The example could not be reused in external project. This patch
creates a library of dtls module to be used externally.

There was also a problem with server stopping accepting connections
when an invalid packet was received. It seems fixed now. the fix is
not perfect, but seems to work and meets my needs.